refactor(net): extract Within trait into unconditionally-compiled module
Move the Within<T> trait and all layer-ordering impls out of the
feature-gated builder module into headers::within (always compiled).

Also:
- Add Within impls for Net/Transport enums (no-op conform) for
  enum-level pattern matching
- Add Within impls for EmbeddedHeaders (after Icmp4/Icmp6) and
  truncated transport types for embedded ICMP payload matching
- Add EmbeddedStart marker for embedded matcher entry position
- Re-export EmbeddedStart publicly from headers module
- Make EmbeddedHeaders.{net, net_ext, transport} pub(super) for
  the pattern matcher (narrowed from pub(crate))

The enum-level Within impls use no-op conform because the concrete
variant is unknown statically.  The builder is safe from these:
HeaderStack::stack additionally requires Blank, which enums do not
implement -- stacking Net or Transport directly is a compile error.
