+++ net/src/headers/builder.rs
@@ -85,8 +85,10 @@ use crate::ipv6::{DestOpts, Fragment, HopByHop, Ipv6, Routing};
 use crate::parse::DeParse;
 use crate::tcp::port::TcpPort;
 use crate::tcp::{Tcp, TruncatedTcp};
+#[cfg(test)]
+use crate::udp::UdpChecksum;
 use crate::udp::port::UdpPort;
-use crate::udp::{Udp, UdpChecksum, UdpEncap};
+use crate::udp::{Udp, UdpEncap};
 use crate::vlan::{Pcp, Vid, Vlan};
 use crate::vxlan::{Vni, Vxlan};
 
@@ -261,18 +263,7 @@ fn fixup_embedded(
     b.build().unwrap_or_else(|_| unreachable!())
 }
 
-/// Declares that `Self` is a valid child of layer `T`.
-///
-/// When `Self` is stacked on a parent `T`, [`conform`](Within::conform)
-/// adjusts structural fields on the parent to be consistent with the child.
-/// For example, `Within<Eth> for Ipv4` sets `EthType::IPV4` on the Ethernet
-/// header.
-///
-/// Conformance is called automatically by [`HeaderStack::stack`] before the parent
-/// is installed into [`Headers`].
-pub trait Within<T> {
-    fn conform(parent: &mut T);
-}
+pub use super::Within;
 
 /// Declares that [`Headers`] can absorb a value of type `T`.
 ///
@@ -525,335 +516,7 @@ where
     );
 }
 
-// Within impls -- valid layer relationships
-impl Within<()> for Eth {
-    fn conform(_parent: &mut ()) {}
-}
-
-impl Within<Eth> for Vlan {
-    fn conform(parent: &mut Eth) {
-        parent.set_ether_type(EthType::VLAN);
-    }
-}
-
-impl Within<Vlan> for Vlan {
-    fn conform(parent: &mut Vlan) {
-        parent.set_inner_ethtype(EthType::VLAN);
-    }
-}
-
-impl Within<Eth> for Ipv4 {
-    fn conform(parent: &mut Eth) {
-        parent.set_ether_type(EthType::IPV4);
-    }
-}
-
-impl Within<Vlan> for Ipv4 {
-    fn conform(parent: &mut Vlan) {
-        parent.set_inner_ethtype(EthType::IPV4);
-    }
-}
-
-impl Within<Eth> for Ipv6 {
-    fn conform(parent: &mut Eth) {
-        parent.set_ether_type(EthType::IPV6);
-    }
-}
-
-impl Within<Vlan> for Ipv6 {
-    fn conform(parent: &mut Vlan) {
-        parent.set_inner_ethtype(EthType::IPV6);
-    }
-}
-
-impl Within<Ipv4> for Tcp {
-    fn conform(parent: &mut Ipv4) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<Ipv6> for Tcp {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<Ipv4> for Udp {
-    fn conform(parent: &mut Ipv4) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Ipv6> for Udp {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Ipv4> for Icmp4 {
-    fn conform(parent: &mut Ipv4) {
-        parent.set_next_header(NextHeader::ICMP);
-    }
-}
-
-impl Within<Ipv6> for Icmp6 {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::ICMP6);
-    }
-}
-
-impl Within<Udp> for Vxlan {
-    fn conform(parent: &mut Udp) {
-        let _ = parent.set_checksum(UdpChecksum::ZERO);
-        parent.set_destination(Vxlan::PORT);
-    }
-}
-
-// ---------------------------------------------------------------------------
-// Within impls for IPv6 extension headers (RFC 8200 section 4.1 ordering)
-// ---------------------------------------------------------------------------
-//
-// Recommended order:
-//   IPv6 -> HopByHop -> DestOpts -> Routing -> Fragment -> AH -> DestOpts -> upper
-//
-// HopByHop MUST immediately follow IPv6 when present.
-// DestOpts may appear in two positions (before Routing, and after AH).
-// The `Within` bounds encode valid parent->child transitions; absence of
-// an impl = compile error = invalid ordering.
-
-// -- HopByHop: only after Ipv6 --
-impl Within<Ipv6> for HopByHop {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::HOP_BY_HOP);
-    }
-}
-
-// -- DestOpts: after Ipv6, HopByHop, Routing, Fragment, or Ipv6Auth --
-// RFC 8200 section 4.1 allows DestOpts in two positions: before Routing
-// (first occurrence) and as the final extension before the upper layer.
-impl Within<Ipv6> for DestOpts {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::DEST_OPTS);
-    }
-}
-
-impl Within<HopByHop> for DestOpts {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::DEST_OPTS);
-    }
-}
-
-impl Within<Routing> for DestOpts {
-    fn conform(parent: &mut Routing) {
-        parent.set_next_header(NextHeader::DEST_OPTS);
-    }
-}
-
-impl Within<Fragment> for DestOpts {
-    fn conform(parent: &mut Fragment) {
-        parent.set_next_header(NextHeader::DEST_OPTS);
-    }
-}
-
-impl Within<Ipv6Auth> for DestOpts {
-    fn conform(parent: &mut Ipv6Auth) {
-        parent.set_next_header(NextHeader::DEST_OPTS);
-    }
-}
-
-// -- Routing: after Ipv6, HopByHop, or DestOpts --
-impl Within<Ipv6> for Routing {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::ROUTING);
-    }
-}
-
-impl Within<HopByHop> for Routing {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::ROUTING);
-    }
-}
-
-impl Within<DestOpts> for Routing {
-    fn conform(parent: &mut DestOpts) {
-        parent.set_next_header(NextHeader::ROUTING);
-    }
-}
-
-// -- Fragment: after Ipv6, HopByHop, DestOpts, or Routing --
-impl Within<Ipv6> for Fragment {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::FRAGMENT);
-    }
-}
-
-impl Within<HopByHop> for Fragment {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::FRAGMENT);
-    }
-}
-
-impl Within<DestOpts> for Fragment {
-    fn conform(parent: &mut DestOpts) {
-        parent.set_next_header(NextHeader::FRAGMENT);
-    }
-}
-
-impl Within<Routing> for Fragment {
-    fn conform(parent: &mut Routing) {
-        parent.set_next_header(NextHeader::FRAGMENT);
-    }
-}
-
-// -- Ipv4Auth: after Ipv4 only --
-impl Within<Ipv4> for Ipv4Auth {
-    fn conform(parent: &mut Ipv4) {
-        parent.set_next_header(NextHeader::AUTH);
-    }
-}
-
-// -- Ipv6Auth: after Ipv6, HopByHop, DestOpts, Routing, or Fragment --
-impl Within<Ipv6> for Ipv6Auth {
-    fn conform(parent: &mut Ipv6) {
-        parent.set_next_header(NextHeader::AUTH);
-    }
-}
-
-impl Within<HopByHop> for Ipv6Auth {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::AUTH);
-    }
-}
-
-impl Within<DestOpts> for Ipv6Auth {
-    fn conform(parent: &mut DestOpts) {
-        parent.set_next_header(NextHeader::AUTH);
-    }
-}
-
-impl Within<Routing> for Ipv6Auth {
-    fn conform(parent: &mut Routing) {
-        parent.set_next_header(NextHeader::AUTH);
-    }
-}
-
-impl Within<Fragment> for Ipv6Auth {
-    fn conform(parent: &mut Fragment) {
-        parent.set_next_header(NextHeader::AUTH);
-    }
-}
-
-// -- Transport after extension headers --
-// Tcp, Udp, Icmp6 can follow any IPv6 extension header.
-// Icmp4 can follow Ipv4Auth (IPv4 context).
-
-impl Within<HopByHop> for Tcp {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<DestOpts> for Tcp {
-    fn conform(parent: &mut DestOpts) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<Routing> for Tcp {
-    fn conform(parent: &mut Routing) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<Fragment> for Tcp {
-    fn conform(parent: &mut Fragment) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<Ipv4Auth> for Tcp {
-    fn conform(parent: &mut Ipv4Auth) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<Ipv6Auth> for Tcp {
-    fn conform(parent: &mut Ipv6Auth) {
-        parent.set_next_header(NextHeader::TCP);
-    }
-}
-
-impl Within<HopByHop> for Udp {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<DestOpts> for Udp {
-    fn conform(parent: &mut DestOpts) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Routing> for Udp {
-    fn conform(parent: &mut Routing) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Fragment> for Udp {
-    fn conform(parent: &mut Fragment) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Ipv4Auth> for Udp {
-    fn conform(parent: &mut Ipv4Auth) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Ipv6Auth> for Udp {
-    fn conform(parent: &mut Ipv6Auth) {
-        parent.set_next_header(NextHeader::UDP);
-    }
-}
-
-impl Within<Ipv4Auth> for Icmp4 {
-    fn conform(parent: &mut Ipv4Auth) {
-        parent.set_next_header(NextHeader::ICMP);
-    }
-}
-
-impl Within<HopByHop> for Icmp6 {
-    fn conform(parent: &mut HopByHop) {
-        parent.set_next_header(NextHeader::ICMP6);
-    }
-}
-
-impl Within<DestOpts> for Icmp6 {
-    fn conform(parent: &mut DestOpts) {
-        parent.set_next_header(NextHeader::ICMP6);
-    }
-}
-
-impl Within<Routing> for Icmp6 {
-    fn conform(parent: &mut Routing) {
-        parent.set_next_header(NextHeader::ICMP6);
-    }
-}
-
-impl Within<Fragment> for Icmp6 {
-    fn conform(parent: &mut Fragment) {
-        parent.set_next_header(NextHeader::ICMP6);
-    }
-}
-
-impl Within<Ipv6Auth> for Icmp6 {
-    fn conform(parent: &mut Ipv6Auth) {
-        parent.set_next_header(NextHeader::ICMP6);
-    }
-}
+// Within impls live in super::within -- re-exported above as `pub use super::Within`.
 
 // Install impls -- how Headers absorbs each layer
